refactor diff editor bottombar
