+++ frontend/src/components/editView/MonacoDiffEditor/MonacoDiffEditor.tsx
@@ -1,11 +1,8 @@
-import { Button, createStyles, makeStyles, useTheme } from '@material-ui/core'
+import { createStyles, makeStyles, useTheme } from '@material-ui/core'
 import { DiffEditor, Monaco } from '@monaco-editor/react'
 import { editor } from 'monaco-editor'
-import React, { useRef, useState } from 'react'
-import useSaveDialog from '../../../hooks/useSaveDialog'
-import LatestCommit from '../LatestCommit'
-import MergeDialog from '../saveDialogs/MergeDialog'
-import ServiceConnected from '../ServiceConnected'
+import React, { useRef } from 'react'
+import ConflictedEditorBottomBar from '../ConflictedBottomBar'
 import useDiffEditor from './useDiffEditor'
 
 interface Props {
@@ -20,13 +17,6 @@ interface Props {
 
 const stylesInUse = makeStyles(() =>
   createStyles({
-    saveGroup: {
-      margin: '10px 20px',
-    },
-    buttonAndStatus: {
-      display: 'flex',
-      alignItems: 'center',
-    },
     title: {
       height: '1rem',
       marginLeft: '20px',
@@ -45,25 +35,9 @@ const MonacoDiffEditor = ({
   currentService,
   updateTheme,
 }: Props) => {
-  const {
-    dialogOpen,
-    dialogError,
-    handleDialogClose,
-    setDialogError,
-    handleDialogOpen,
-  } = useSaveDialog()
-
-  const [waitingToMerge, setWaitingToMerge] = useState(false)
-
   const classes = stylesInUse()
 
-  const {
-    setupCodeLens,
-    mergeConflictExists,
-    saveMergeEdit,
-    modifiedContent,
-    mutationMergeLoading,
-  } = useDiffEditor(original, cloneUrl)
+  const { setupCodeLens, modifiedContent } = useDiffEditor(original, cloneUrl)
 
   const theme = useTheme()
 
@@ -76,40 +50,6 @@ const MonacoDiffEditor = ({
     editorRef.current = editor
 
     setupCodeLens(editor, monaco)
-  }
-
-  const handleDialogSubmit = async (newCommitMessage: string) => {
-    if (editorRef.current) {
-      try {
-        setWaitingToMerge(true)
-        await saveMergeEdit({
-          variables: {
-            files: [
-              {
-                name: filename,
-                content: editorRef.current.getModifiedEditor().getValue(),
-              },
-            ],
-            commitMessage: newCommitMessage,
-          },
-        })
-        handleDialogClose()
-        setDialogError(undefined)
-      } catch (error) {
-        const dialogError = {
-          title: `An error occurred while merging`,
-          message: '',
-        }
-
-        if (error instanceof Error) {
-          dialogError.message = `More info: ${error.message}`
-        }
-
-        setDialogError(dialogError)
-      } finally {
-        setWaitingToMerge(false)
-      }
-    }
   }
 
   const options: editor.IDiffEditorConstructionOptions = {
@@ -134,29 +74,13 @@ const MonacoDiffEditor = ({
         // Updating the theme here so we override things set by <Editor>
         updateTheme()
       }
-      <MergeDialog
-        open={dialogOpen}
-        handleClose={handleDialogClose}
-        handleSubmit={handleDialogSubmit}
+      <ConflictedEditorBottomBar
+        cloneUrl={cloneUrl}
+        original={original}
         currentBranch={currentBranch}
-        error={dialogError}
-        waitingToMerge={waitingToMerge}
+        currentService={currentService}
+        commitMessage={commitMessage}
       />
-
-      <div className={classes.saveGroup}>
-        <div className={classes.buttonAndStatus}>
-          <Button
-            color="primary"
-            variant="contained"
-            disabled={mutationMergeLoading || mergeConflictExists}
-            onClick={handleDialogOpen}
-          >
-            Merge
-          </Button>
-          <ServiceConnected service={currentService} />
-          <LatestCommit commitMessage={commitMessage} />
-        </div>
-      </div>
     </div>
   )
 }
